fix: make initial migration script compatible with PostGIS
