+++ alembic/versions/fb089dbbfc43_initial_migration.py
@@ -1,0 +1,627 @@
+"""Initial migration
+
+Revision ID: fb089dbbfc43
+Revises: 
+Create Date: 2025-07-22 11:12:10.398798
+
+"""
+
+from typing import Sequence, Union
+
+from alembic import op
+import geoalchemy2
+import sqlalchemy as sa
+import sqlalchemy_utils
+
+
+# revision identifiers, used by Alembic.
+revision: str = "fb089dbbfc43"
+down_revision: Union[str, Sequence[str], None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table(
+        "asset",
+        sa.Column("filename", sa.String(), nullable=True),
+        sa.Column("storage_service", sa.String(), nullable=True),
+        sa.Column("storage_path", sa.String(), nullable=True),
+        sa.Column("mime_type", sa.String(), nullable=True),
+        sa.Column("size", sa.Integer(), nullable=True),
+        sa.Column(
+            "search_vector",
+            sqlalchemy_utils.types.ts_vector.TSVectorType(),
+            nullable=True,
+        ),
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column(
+            "created_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(
+        "ix_asset_search_vector",
+        "asset",
+        ["search_vector"],
+        unique=False,
+        postgresql_using="gin",
+    )
+    op.create_table(
+        "lexicon_category",
+        sa.Column("name", sa.String(length=100), nullable=False),
+        sa.Column("description", sa.String(length=255), nullable=True),
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column(
+            "created_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
+        ),
+        sa.PrimaryKeyConstraint("id"),
+        sa.UniqueConstraint("name"),
+    )
+    op.create_table(
+        "lexicon_term",
+        sa.Column("term", sa.String(length=100), nullable=False),
+        sa.Column("definition", sa.String(length=255), nullable=False),
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column(
+            "created_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
+        ),
+        sa.PrimaryKeyConstraint("id"),
+        sa.UniqueConstraint("term"),
+    )
+    op.create_table(
+        "pub_author",
+        sa.Column("name", sa.String(), nullable=False),
+        sa.Column("affiliation", sa.String(), nullable=True),
+        sa.Column(
+            "search_vector",
+            sqlalchemy_utils.types.ts_vector.TSVectorType(),
+            nullable=True,
+        ),
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column(
+            "created_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(
+        "ix_pub_author_search_vector",
+        "pub_author",
+        ["search_vector"],
+        unique=False,
+        postgresql_using="gin",
+    )
+    op.create_table(
+        "sensor",
+        sa.Column("name", sa.String(length=255), nullable=False),
+        sa.Column("model", sa.String(length=50), nullable=True),
+        sa.Column("serial_no", sa.String(length=50), nullable=True),
+        sa.Column("date_installed", sa.DateTime(), nullable=True),
+        sa.Column("date_removed", sa.DateTime(), nullable=True),
+        sa.Column("recording_interval", sa.Integer(), nullable=True),
+        sa.Column("notes", sa.String(length=50), nullable=True),
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column(
+            "created_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_table(
+        "user",
+        sa.Column("id", sa.Integer(), nullable=False),
+        sa.Column("username", sa.String(length=255), nullable=False),
+        sa.Column("password", sa.String(length=255), nullable=False),
+        sa.Column("is_superuser", sa.Boolean(), nullable=False),
+        sa.Column("is_active", sa.Boolean(), nullable=False),
+        sa.Column("avatar_url", sa.Text(), nullable=True),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_table(
+        "contact",
+        sa.Column("name", sa.String(length=100), nullable=False),
+        sa.Column("role", sa.String(length=100), nullable=False),
+        sa.Column(
+            "search_vector",
+            sqlalchemy_utils.types.ts_vector.TSVectorType(),
+            nullable=True,
+        ),
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column(
+            "created_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
+        ),
+        sa.ForeignKeyConstraint(
+            ["role"],
+            ["lexicon_term.term"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(
+        "ix_contact_search_vector",
+        "contact",
+        ["search_vector"],
+        unique=False,
+        postgresql_using="gin",
+    )
+    op.create_table(
+        "geochronology_age",
+        sa.Column("location_id", sa.Integer(), nullable=False),
+        sa.Column("age", sa.Float(), nullable=False),
+        sa.Column("age_error", sa.Float(), nullable=True),
+        sa.Column("method", sa.String(length=100), nullable=True),
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column(
+            "created_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
+        ),
+        sa.ForeignKeyConstraint(
+            ["method"],
+            ["lexicon_term.term"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_table(
+        "groundwater_level_sensor",
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column(
+            "created_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
+        ),
+        sa.Column("sensor_id", sa.Integer(), nullable=False),
+        sa.ForeignKeyConstraint(["sensor_id"], ["sensor.id"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("id"),
+        sa.UniqueConstraint("sensor_id"),
+    )
+    op.create_table(
+        "lexicon_term_category_association",
+        sa.Column("lexicon_term", sa.String(length=100), nullable=False),
+        sa.Column("category_name", sa.String(length=255), nullable=False),
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column(
+            "created_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
+        ),
+        sa.ForeignKeyConstraint(
+            ["category_name"], ["lexicon_category.name"], ondelete="CASCADE"
+        ),
+        sa.ForeignKeyConstraint(
+            ["lexicon_term"], ["lexicon_term.term"], ondelete="CASCADE"
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_table(
+        "lexicon_triple",
+        sa.Column("subject", sa.String(length=100), nullable=False),
+        sa.Column("predicate", sa.String(length=100), nullable=False),
+        sa.Column("object_", sa.String(length=100), nullable=False),
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column(
+            "created_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
+        ),
+        sa.ForeignKeyConstraint(["object_"], ["lexicon_term.term"], ondelete="CASCADE"),
+        sa.ForeignKeyConstraint(["subject"], ["lexicon_term.term"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_table(
+        "location",
+        sa.Column("notes", sa.Text(), nullable=True),
+        sa.Column(
+            "point",
+            geoalchemy2.types.Geometry(
+                geometry_type="POINT",
+                srid=4326,
+                from_text="ST_GeomFromEWKT",
+                name="geometry",
+                nullable=False,
+            ),
+            nullable=False,
+        ),
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column(
+            "created_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
+        ),
+        sa.Column("release_status", sa.String(length=100), nullable=True),
+        sa.ForeignKeyConstraint(
+            ["release_status"],
+            ["lexicon_term.term"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(
+        "idx_location_point",
+        "location",
+        ["point"],
+        unique=False,
+        postgresql_using="gist",
+    )
+    op.create_table(
+        "publication",
+        sa.Column("title", sa.Text(), nullable=False),
+        sa.Column("abstract", sa.Text(), nullable=True),
+        sa.Column("doi", sa.String(), nullable=True),
+        sa.Column("year", sa.Integer(), nullable=True),
+        sa.Column("publisher", sa.String(), nullable=True),
+        sa.Column("url", sa.String(), nullable=True),
+        sa.Column("publication_type", sa.String(length=100), nullable=False),
+        sa.Column(
+            "search_vector",
+            sqlalchemy_utils.types.ts_vector.TSVectorType(),
+            nullable=True,
+        ),
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column(
+            "created_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
+        ),
+        sa.ForeignKeyConstraint(
+            ["publication_type"],
+            ["lexicon_term.term"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+        sa.UniqueConstraint("doi"),
+    )
+    op.create_index(
+        "ix_publication_search_vector",
+        "publication",
+        ["search_vector"],
+        unique=False,
+        postgresql_using="gin",
+    )
+    op.create_table(
+        "thing",
+        sa.Column("name", sa.String(length=255), nullable=False),
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column(
+            "created_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
+        ),
+        sa.Column("release_status", sa.String(length=100), nullable=True),
+        sa.ForeignKeyConstraint(
+            ["release_status"],
+            ["lexicon_term.term"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_table(
+        "address",
+        sa.Column("contact_id", sa.Integer(), nullable=False),
+        sa.Column("address_line_1", sa.String(length=255), nullable=False),
+        sa.Column("address_line_2", sa.String(length=255), nullable=True),
+        sa.Column("city", sa.String(length=100), nullable=False),
+        sa.Column("state", sa.String(length=50), nullable=False),
+        sa.Column("postal_code", sa.String(length=20), nullable=False),
+        sa.Column("country", sa.String(length=100), nullable=True),
+        sa.Column("address_type", sa.String(length=100), nullable=True),
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column(
+            "created_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
+        ),
+        sa.ForeignKeyConstraint(
+            ["address_type"],
+            ["lexicon_term.term"],
+        ),
+        sa.ForeignKeyConstraint(["contact_id"], ["contact.id"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_table(
+        "asset_thing_association",
+        sa.Column("asset_id", sa.Integer(), nullable=False),
+        sa.Column("thing_id", sa.Integer(), nullable=False),
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column(
+            "created_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
+        ),
+        sa.ForeignKeyConstraint(["asset_id"], ["asset.id"], ondelete="CASCADE"),
+        sa.ForeignKeyConstraint(["thing_id"], ["thing.id"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_table(
+        "email",
+        sa.Column("contact_id", sa.Integer(), nullable=False),
+        sa.Column("email", sa.String(length=100), nullable=False),
+        sa.Column("email_type", sa.String(length=100), nullable=True),
+        sa.Column(
+            "search_vector",
+            sqlalchemy_utils.types.ts_vector.TSVectorType(),
+            nullable=True,
+        ),
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column(
+            "created_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
+        ),
+        sa.ForeignKeyConstraint(["contact_id"], ["contact.id"], ondelete="CASCADE"),
+        sa.ForeignKeyConstraint(
+            ["email_type"],
+            ["lexicon_term.term"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(
+        "ix_email_search_vector",
+        "email",
+        ["search_vector"],
+        unique=False,
+        postgresql_using="gin",
+    )
+    op.create_table(
+        "location_thing_association",
+        sa.Column("location_id", sa.Integer(), nullable=False),
+        sa.Column("thing_id", sa.Integer(), nullable=False),
+        sa.Column(
+            "effective_start",
+            sa.DateTime(),
+            server_default=sa.text("now()"),
+            nullable=False,
+        ),
+        sa.Column("effective_end", sa.DateTime(), nullable=True),
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column(
+            "created_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
+        ),
+        sa.ForeignKeyConstraint(["location_id"], ["location.id"], ondelete="CASCADE"),
+        sa.ForeignKeyConstraint(["thing_id"], ["thing.id"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("location_id", "thing_id", "id"),
+    )
+    op.create_table(
+        "phone",
+        sa.Column("contact_id", sa.Integer(), nullable=False),
+        sa.Column("phone_number", sa.String(length=20), nullable=False),
+        sa.Column("phone_type", sa.String(length=100), nullable=True),
+        sa.Column(
+            "search_vector",
+            sqlalchemy_utils.types.ts_vector.TSVectorType(),
+            nullable=True,
+        ),
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column(
+            "created_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
+        ),
+        sa.ForeignKeyConstraint(["contact_id"], ["contact.id"], ondelete="CASCADE"),
+        sa.ForeignKeyConstraint(
+            ["phone_type"],
+            ["lexicon_term.term"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(
+        "ix_phone_search_vector",
+        "phone",
+        ["search_vector"],
+        unique=False,
+        postgresql_using="gin",
+    )
+    op.create_table(
+        "pub_author_contact_association",
+        sa.Column("author_id", sa.Integer(), nullable=False),
+        sa.Column("contact_id", sa.Integer(), nullable=False),
+        sa.Column(
+            "created_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
+        ),
+        sa.ForeignKeyConstraint(["author_id"], ["pub_author.id"], ondelete="CASCADE"),
+        sa.ForeignKeyConstraint(["contact_id"], ["contact.id"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("author_id", "contact_id"),
+    )
+    op.create_table(
+        "pub_author_publication_association",
+        sa.Column("publication_id", sa.Integer(), nullable=False),
+        sa.Column("author_id", sa.Integer(), nullable=False),
+        sa.Column("author_order", sa.Integer(), nullable=False),
+        sa.Column(
+            "created_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
+        ),
+        sa.ForeignKeyConstraint(["author_id"], ["pub_author.id"], ondelete="CASCADE"),
+        sa.ForeignKeyConstraint(
+            ["publication_id"], ["publication.id"], ondelete="CASCADE"
+        ),
+        sa.PrimaryKeyConstraint("publication_id", "author_id"),
+    )
+    op.create_table(
+        "series",
+        sa.Column("observed_property", sa.String(length=100), nullable=False),
+        sa.Column("unit", sa.String(length=100), nullable=False),
+        sa.Column("name", sa.String(length=255), nullable=False),
+        sa.Column("description", sa.Text(), nullable=True),
+        sa.Column("sensor_id", sa.Integer(), nullable=True),
+        sa.Column("thing_id", sa.Integer(), nullable=False),
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column(
+            "created_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
+        ),
+        sa.Column("release_status", sa.String(length=100), nullable=True),
+        sa.ForeignKeyConstraint(
+            ["observed_property"],
+            ["lexicon_term.term"],
+        ),
+        sa.ForeignKeyConstraint(
+            ["release_status"],
+            ["lexicon_term.term"],
+        ),
+        sa.ForeignKeyConstraint(["sensor_id"], ["sensor.id"], ondelete="CASCADE"),
+        sa.ForeignKeyConstraint(["thing_id"], ["thing.id"], ondelete="CASCADE"),
+        sa.ForeignKeyConstraint(
+            ["unit"],
+            ["lexicon_term.term"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_table(
+        "spring_thing",
+        sa.Column("description", sa.String(length=255), nullable=True),
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column(
+            "created_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
+        ),
+        sa.Column("thing_id", sa.Integer(), nullable=False),
+        sa.ForeignKeyConstraint(["thing_id"], ["thing.id"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("id"),
+        sa.UniqueConstraint("thing_id"),
+    )
+    op.create_table(
+        "thing_contact_association",
+        sa.Column("thing_id", sa.Integer(), nullable=False),
+        sa.Column("contact_id", sa.Integer(), nullable=False),
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column(
+            "created_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
+        ),
+        sa.ForeignKeyConstraint(
+            ["contact_id"],
+            ["contact.id"],
+        ),
+        sa.ForeignKeyConstraint(
+            ["thing_id"],
+            ["thing.id"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_table(
+        "thing_id_link",
+        sa.Column("thing_id", sa.Integer(), nullable=True),
+        sa.Column("relation", sa.String(length=100), nullable=False),
+        sa.Column("alternate_id", sa.String(length=100), nullable=False),
+        sa.Column("alternate_organization", sa.String(length=100), nullable=False),
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column(
+            "created_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
+        ),
+        sa.ForeignKeyConstraint(
+            ["alternate_organization"],
+            ["lexicon_term.term"],
+        ),
+        sa.ForeignKeyConstraint(
+            ["relation"],
+            ["lexicon_term.term"],
+        ),
+        sa.ForeignKeyConstraint(["thing_id"], ["thing.id"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_table(
+        "well_thing",
+        sa.Column("well_depth", sa.Float(), nullable=True),
+        sa.Column("hole_depth", sa.Float(), nullable=True),
+        sa.Column("well_type", sa.String(length=100), nullable=True),
+        sa.Column("casing_diameter", sa.Float(), nullable=True),
+        sa.Column("casing_depth", sa.Float(), nullable=True),
+        sa.Column("casing_description", sa.String(length=50), nullable=True),
+        sa.Column("construction_notes", sa.String(length=250), nullable=True),
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column(
+            "created_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
+        ),
+        sa.Column("thing_id", sa.Integer(), nullable=False),
+        sa.ForeignKeyConstraint(["thing_id"], ["thing.id"], ondelete="CASCADE"),
+        sa.ForeignKeyConstraint(
+            ["well_type"],
+            ["lexicon_term.term"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+        sa.UniqueConstraint("thing_id"),
+    )
+    op.create_table(
+        "collaborative_network_well",
+        sa.Column("actively_monitored", sa.Boolean(), nullable=False),
+        sa.Column("well_id", sa.Integer(), nullable=False),
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column(
+            "created_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
+        ),
+        sa.ForeignKeyConstraint(["well_id"], ["well_thing.id"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_table(
+        "geochemical_series",
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column(
+            "created_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
+        ),
+        sa.Column("series_id", sa.Integer(), nullable=False),
+        sa.ForeignKeyConstraint(["series_id"], ["series.id"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("id"),
+        sa.UniqueConstraint("series_id"),
+    )
+    op.create_table(
+        "geothermal_series",
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column(
+            "created_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
+        ),
+        sa.Column("series_id", sa.Integer(), nullable=False),
+        sa.ForeignKeyConstraint(["series_id"], ["series.id"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("id"),
+        sa.UniqueConstraint("series_id"),
+    )
+    op.create_table(
+        "groundwater_level_series",
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column(
+            "created_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
+        ),
+        sa.Column("series_id", sa.Integer(), nullable=False),
+        sa.ForeignKeyConstraint(["series_id"], ["series.id"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("id"),
+        sa.UniqueConstraint("series_id"),
+    )
+    op.create_table(
+        "well_screen",
+        sa.Column("well_id", sa.Integer(), nullable=False),
+        sa.Column("screen_depth_top", sa.Float(), nullable=False),
+        sa.Column("screen_depth_bottom", sa.Float(), nullable=False),
+        sa.Column("screen_type", sa.String(length=100), nullable=True),
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column(
+            "created_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
+        ),
+        sa.ForeignKeyConstraint(
+            ["screen_type"],
+            ["lexicon_term.term"],
+        ),
+        sa.ForeignKeyConstraint(["well_id"], ["well_thing.id"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table("well_screen")
+    op.drop_table("groundwater_level_series")
+    op.drop_table("geothermal_series")
+    op.drop_table("geochemical_series")
+    op.drop_table("collaborative_network_well")
+    op.drop_table("well_thing")
+    op.drop_table("thing_id_link")
+    op.drop_table("thing_contact_association")
+    op.drop_table("spring_thing")
+    op.drop_table("series")
+    op.drop_table("pub_author_publication_association")
+    op.drop_table("pub_author_contact_association")
+    op.drop_index("ix_phone_search_vector", table_name="phone", postgresql_using="gin")
+    op.drop_table("phone")
+    op.drop_table("location_thing_association")
+    op.drop_index("ix_email_search_vector", table_name="email", postgresql_using="gin")
+    op.drop_table("email")
+    op.drop_table("asset_thing_association")
+    op.drop_table("address")
+    op.drop_table("thing")
+    op.drop_index(
+        "ix_publication_search_vector", table_name="publication", postgresql_using="gin"
+    )
+    op.drop_table("publication")
+    op.drop_index("idx_location_point", table_name="location", postgresql_using="gist")
+    op.drop_table("location")
+    op.drop_table("lexicon_triple")
+    op.drop_table("lexicon_term_category_association")
+    op.drop_table("groundwater_level_sensor")
+    op.drop_table("geochronology_age")
+    op.drop_index(
+        "ix_contact_search_vector", table_name="contact", postgresql_using="gin"
+    )
+    op.drop_table("contact")
+    op.drop_table("user")
+    op.drop_table("sensor")
+    op.drop_index(
+        "ix_pub_author_search_vector", table_name="pub_author", postgresql_using="gin"
+    )
+    op.drop_table("pub_author")
+    op.drop_table("lexicon_term")
+    op.drop_table("lexicon_category")
+    op.drop_index("ix_asset_search_vector", table_name="asset", postgresql_using="gin")
+    op.drop_table("asset")
+    # ### end Alembic commands ###
